+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "Coursecare",
   "version": "1.0.0",
-  "description": "a site for collaborative learning",
+  "description": "A site for collaborative learning",
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
