+++ src/routes/index.js
@@ -6,19 +6,19 @@
 import bcrypt from 'bcryptjs'
 // import tutor from './tutor'
 // import student from './student'
-import { index, registerUser, register, dashboard, profile, addCourse, registerCourse, courseList, searchCourseTemplate } from './../controllers'
-import { whiteboard, discussion, addCourseTemplate, registerCourseTemplate, comingSoon } from './../controllers/ajax'
+import { index, registerUser, register, dashboard, profile } from './../controllers'
+import { whiteboardTemplate, discussion, comingSoon } from './../controllers/ajax'
+import { addCourseTemplate, registerCourseTemplate, searchCourseTemplate, courseListTemplate, addCourse, registerCourse } from '../controllers/course'
 
 module.exports = function (app, passport) {
 
     app.get('/', index)
 
 
-    // facebook oauth
-
+    // oauth
     app.get('/auth/facebook', passport.authenticate('facebook'))
-
     app.get('/auth/google', passport.authenticate('google', { scope: ['https://www.googleapis.com/auth/plus.login', 'email'] }))
+
 
     app.get('/auth/google/callback', function (req, res, next) {
         passport.authenticate('google', function (err, user, info) {
@@ -33,7 +33,6 @@ module.exports = function (app, passport) {
                 })
         })(req, res, next)
     })
-
     app.get('/auth/facebook/callback', function (req, res, next) {
         passport.authenticate('facebook', function (err, user, info) {
             console.log("callack facebook")
@@ -48,22 +47,24 @@ module.exports = function (app, passport) {
         })(req, res, next)
     })
 
-    app.get('/whiteboard', whiteboard)
+    app.get('/whiteboard', whiteboardTemplate)
     app.get('/qanda', comingSoon)
     app.get('/discussion', discussion)
+
+    // courses
     app.get('/course/add', addCourseTemplate)
     app.get('/course/search', searchCourseTemplate)
     app.get('/course/register', registerCourseTemplate)
     app.get('/course/register/:id', registerCourseTemplate)
+    app.get('/course/list', courseListTemplate)
 
     app.post('/course/register', registerCourse)
     app.post('/course/add', addCourse)
-    app.get('/course/list', courseList)
 
     app.get('/:role/register', register)
     app.post('/:role/register', registerUser)
-    app.get('/:role/dashboard', dashboard)
     app.get('/:role/profile', profile)
+    app.get('/:role/dashboard', dashboard)
     app.get('/:role/dashboard/:id', dashboard)
 
     app.post('/:role/login', function (req, res, next) {
@@ -80,8 +81,6 @@ module.exports = function (app, passport) {
     })
 
 
-    // app.use('/tutor', tutor)
-    // app.use('/student', student)
 
     app.get('/logout', function (req, res) {
         let role = req.user.role;
