+++ src/views/partials/feed.handlebars
@@ -1,47 +1,12 @@
 <div class="pusher">
     <div class="ui text disabled loader" id="loader">Loading</div>
-    <!-- <div class="ui basic segment"> -->
     <p class="height-same"></p>
     <div class="ui padded basic segment main-screen">
 {{#if profile}}
      <h3 class="ui header">User Profile</h3>
-     <p>Coming Soom :)</p>
+     <p>Coming Soon :)</p>
     {{else}}
     {{> whiteboard}}
 {{/if}}
-        {{!-- <h3 class="ui header">Application Content</h3>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p>
-        <p>Application Content</p> --}}
     </div>
-    <!-- </div> -->
 </div>
